+++ foliant/preprocessors/testrail/testrail.py
@@ -21,23 +21,34 @@ class Preprocessor(BasePreprocessor):
         'filename': 'test_cases.md',
         'rewrite_src_file': False,
         'template_folder': 'case_templates',
-        'platform_id': 0,
-        'platforms': 'smarttv, androidtv, appletv, web',
         'section_header': 'Программа испытаний',
         'std_table_header': 'Таблица прохождения испытаний',
-        'std_table_column_headers': '№; ID; Название; Успешно; Комментарий',
+        'std_table_column_headers': '№; Приоритет; Платформа; ID; Название; Успешно; Комментарий',
         'suite_ids': set(),
         'section_ids': set(),
+        'exclude_suite_ids': set(),
+        'exclude_section_ids': set(),
+        'exclude_case_ids': set(),
         'add_suite_headers': True,
         'add_section_headers': True,
-        'add_cases_without_platform': True,
-        'add_unpublished_cases': True,
-        'add_case_id_to_case_name': False,
+        'add_case_id_to_case_header': False,
+        'add_case_id_to_std_table': False,
         'add_std_table': True,
         'resolve_urls': False,
         'screenshots_url': '',
         'screenshots_ext': '.png',
         'print_case_structure': False,
+        'multi_param_name': '',
+        'multi_param_select': '',
+        'multi_param_select_type': 'any',
+        'add_cases_without_multi_param': True,
+        'add_multi_param_to_case_header': False,
+        'add_multi_param_to_std_table': False,
+        'checkbox_param_name': '',
+        'checkbox_param_select_type': 'checked',
+        'choose_priorities': '',
+        'add_priority_to_case_header': False,
+        'add_priority_to_std_table': False,
     }
 
 
@@ -54,7 +65,7 @@ class Preprocessor(BasePreprocessor):
         self._section_header = self.options['section_header']
 
         self._std_table_header = self.options['std_table_header']
-        self._std_table_column_headers = self.options['std_table_column_headers'].replace(' ', '').split(';')
+        self._std_table_column_headers = list(map(str.strip, list(self.options['std_table_column_headers'].split(';'))))
 
         self._testrail_url = self.options['testrail_url']
         self._testrail_login = self.options['testrail_login']
@@ -68,27 +79,42 @@ class Preprocessor(BasePreprocessor):
 
         self._suite_ids = set()
         self._suite_ids = self._parse_ids_options(self._suite_ids, self.options['suite_ids'])
+        self._exclude_suite_ids = set()
+        self._exclude_suite_ids = self._parse_ids_options(self._exclude_suite_ids, self.options['exclude_suite_ids'])
 
         self._section_ids = set()
         self._section_ids = self._parse_ids_options(self._section_ids, self.options['section_ids'])
+        self._exclude_section_ids = set()
+        self._exclude_section_ids = self._parse_ids_options(self._exclude_section_ids, self.options['exclude_section_ids'])
 
-        self._platform_id = self.options['platform_id']
-        self._platforms = self.options['platforms'].replace(' ', '').split(',')
-        if not self._platform_id == self.defaults['platform_id']:
-            self._platform_name = self._platforms[self._platform_id-1]
-        else:
-            self._platform_name = ''
+        self._exclude_case_ids = set()
+        self._exclude_case_ids = self._parse_ids_options(self._exclude_case_ids, self.options['exclude_case_ids'])
+
+        self._params = {
+            'multi_param_name': self.options['multi_param_name'],
+            'multi_param_sys_name': 'custom_' + self.options['multi_param_name'],
+            'multi_param_select': list(map(str.lower, map(str.strip, list(self.options['multi_param_select'].split(','))))),
+            'multi_param_select_type': self.options['multi_param_select_type'],
+            'add_cases_without_multi_param': self.options['add_cases_without_multi_param'],
+            'checkbox_param_name': self.options['checkbox_param_name'],
+            'checkbox_param_sys_name': 'custom_' + self.options['checkbox_param_name'],
+            'checkbox_param_select_type': self.options['checkbox_param_select_type'],
+            'choose_priorities': list(map(str.lower, map(str.strip, list(self.options['choose_priorities'].split(','))))),
+            'add_multi_param_to_case_header': self.options['add_multi_param_to_case_header'],
+            'add_multi_param_to_std_table': self.options['add_multi_param_to_std_table'],
+            'add_priority_to_case_header': self.options['add_priority_to_case_header'],
+            'add_priority_to_std_table': self.options['add_priority_to_std_table'],
+            'add_case_id_to_case_header': self.options['add_case_id_to_case_header'],
+            'add_case_id_to_std_table': self.options['add_case_id_to_std_table']
+        }
 
         self._add_suite_headers = self.options['add_suite_headers']
         self._add_section_headers = self.options['add_section_headers']
-        self._add_cases_without_platform = self.options['add_cases_without_platform']
-        self._add_unpublished_cases = self.options['add_unpublished_cases']
-        self._add_case_id_to_case_name = self.options['add_case_id_to_case_name']
         self._add_std_table = self.options['add_std_table']
 
         self._resolve_urls = self.options['resolve_urls']
-        if not self._platform_id == self.defaults['platform_id']:
-            self._screenshots_url = '/'.join((self.options['screenshots_url'], 'raw/master/images', self._platform_name, ''))
+        if self._params['multi_param_select'][0] != '':
+            self._screenshots_url = '/'.join((self.options['screenshots_url'], 'raw/master/images', self._params['multi_param_select'][0], ''))
         else:
             self._screenshots_url = '/'.join((self.options['screenshots_url'], 'raw/master/images/'))
         self._screenshots_ext = self.options['screenshots_ext']
@@ -104,6 +130,7 @@ class Preprocessor(BasePreprocessor):
         self._template_folder = self.options['template_folder']
         if self._template_folder == self.defaults['template_folder'] and not os.path.exists(self.project_path / self.defaults['template_folder']):
             copytree(resource_filename(__name__, 'case_templates'), self.project_path / self.defaults['template_folder'])
+
 
     def _parse_ids_options(self, variable, option):
         if option:
@@ -127,13 +154,14 @@ class Preprocessor(BasePreprocessor):
                         if section['id'] in self._section_ids:
                             self._suite_ids.add(suite['id'])
                             continue
+        self._suite_ids -= self._exclude_suite_ids
+
         if not self._section_ids:
             for suite in project_suites:
                 if suite['id'] in self._suite_ids:
                     suite_sections = self._client.send_get('get_sections/%s&suite_id=%s' % (self._project_id, suite['id']))
                     for section in suite_sections:    
                         self._section_ids.add(section['id'])
-
         next_iteration = True  # collect all child subsections for sections specified
         while next_iteration:
             next_iteration = False
@@ -144,6 +172,7 @@ class Preprocessor(BasePreprocessor):
                         if section['id'] not in self._section_ids and section['parent_id'] in self._section_ids:    
                             self._section_ids.add(section['id'])
                             next_iteration = True
+        self._section_ids -= self._exclude_section_ids
 
 
     def _collect_cases(self, project_suites):
@@ -155,10 +184,16 @@ class Preprocessor(BasePreprocessor):
                     self._test_cases.append('## %s\n\n' % suite['name'])
                     suite['name'] = ''.join(('**', suite['name'].upper(), '**'))
 
-                    if self._add_case_id_to_case_name:
-                        self._std_table.append((' | '.join(('', ' ', ' ', suite['name'], ' ', ' ', ' '))).strip())
-                    else:
-                        self._std_table.append((' | '.join(('', ' ', suite['name'], ' ', ' ', ' '))).strip())
+                    table_row = '|   | '
+                    if self._params['add_priority_to_std_table']:
+                        table_row += '  | '
+                    if self._params['add_multi_param_to_std_table']:
+                        table_row += '  | '
+                    if self._params['add_case_id_to_std_table']:
+                        table_row += '  | '
+                    table_row += suite['name'] + ' |   |   |'
+
+                    self._std_table.append(table_row)
 
                     if suite['description']:
                         self._test_cases.append('%s\n\n' % suite['description'])
@@ -190,10 +225,17 @@ class Preprocessor(BasePreprocessor):
                                 section_name))
 
                 section_name = ''.join(('**', section_name, '**'))
-                if self._add_case_id_to_case_name:
-                    self._std_table.append((' | '.join(('', ' ', ' ', section_name, ' ', ' ', ' '))).strip())
-                else:
-                    self._std_table.append((' | '.join(('', ' ', section_name, ' ', ' ', ' '))).strip())
+
+                table_row = '|   | '
+                if self._params['add_priority_to_std_table']:
+                    table_row += '  | '
+                if self._params['add_multi_param_to_std_table']:
+                    table_row += '  | '
+                if self._params['add_case_id_to_std_table']:
+                    table_row += '  | '
+                table_row += section_name + ' |   |   |'
+
+                self._std_table.append(table_row)
 
             if section['id'] in self._section_ids:  # This condition is checked not earlier to save parent chapter headers
 
@@ -206,32 +248,83 @@ class Preprocessor(BasePreprocessor):
                 self._collect_case_data(suite_id, section['id'], title_level_up)
 
 
+    def _if_take_case(self, case):
+
+        multi_param_cond = True
+        checkbox_param_cond = True
+        priority_cond = True
+        exclude_cond = True
+
+        if self._params['multi_param_sys_name'] in case.keys():
+            if self._params['multi_param_sys_name'] not in case.keys() and not self._params['add_cases_without_multi_param']:
+                    multi_param_cond = False
+            elif self._params['multi_param_select_type'] == 'any' and not (set(case[self._params['multi_param_sys_name']]) & self._params['multi_param_matches']):
+                multi_param_cond = False
+            elif self._params['multi_param_select_type'] == 'only' and not (set(case[self._params['multi_param_sys_name']]) <= self._params['multi_param_matches']):
+                multi_param_cond = False
+            elif self._params['multi_param_select_type'] == 'all' and not (self._params['multi_param_matches'] <= set(case[self._params['multi_param_sys_name']])):
+                multi_param_cond = False
+            elif self._params['multi_param_select_type'] == 'match' and not (set(case[self._params['multi_param_sys_name']]) == self._params['multi_param_matches']):
+                multi_param_cond = False
+
+        if self._params['checkbox_param_sys_name'] in case.keys():
+            if self._params['checkbox_param_select_type'] == 'checked' and not case[self._params['checkbox_param_sys_name']]:
+                checkbox_param_cond = False
+            if self._params['checkbox_param_select_type'] == 'unchecked' and case[self._params['checkbox_param_sys_name']]:
+                checkbox_param_cond = False
+
+        if self._params['choose_priorities'][0] != '' and case['priority_id'] not in self._params['priority_matches']:
+            priority_cond = False
+
+        if case['id'] in self._exclude_case_ids:
+            exclude_cond = False
+
+        take_case = multi_param_cond and checkbox_param_cond and priority_cond and exclude_cond
+
+        return take_case
+
+
     def _collect_case_data(self, suite_id, section_id, title_level_up):
         section_cases = self._client.send_get(
             'get_cases/%s&suite_id=%s&section_id=%s' %
             (self._project_id, suite_id, section_id))
 
         for case in section_cases:
+            take_case = self._if_take_case(case)
 
-            if 'custom_prj_type' not in case.keys():  # Collect cases without platform assigned or not
-                if self._add_cases_without_platform:
-                    case.update({'custom_prj_type': [self._platform_id]})
-                else:
-                    case.update({'custom_prj_type': [0]})
-
-            if 'custom_tp' not in case.keys():  # If test-case template has no TP field (test-case published), test-cases will be collected on 'add_unpublished_cases' value basis
-                case.update({'custom_tp': None})
-
-            if (case['custom_tp'] or self._add_unpublished_cases) and self._platform_id in case['custom_prj_type']:  # Second codition is for test-cases without platform assigned
-
+            if take_case:
                 self._case_counter += 1
 
-                if self._add_case_id_to_case_name:
-                    self._std_table.append((' | '.join(('', str(self._case_counter), str(case['id']), case['title'], ' ', ' ', ' '))).strip())
-                    self._test_cases.append('##%s %s (C%s)\n\n' % (title_level_up, case['title'], case['id']))
-                else:
-                    self._std_table.append((' | '.join(('', str(self._case_counter), case['title'], ' ', ' ', ' '))).strip())
-                    self._test_cases.append('##%s %s\n\n' % (title_level_up, case['title']))
+                table_row = '| ' + str(self._case_counter) + ' | '
+                header_ending = ''
+                header_ending_list = []
+
+                if self._params['add_priority_to_std_table']:
+                    table_row += self._priorities[case['priority_id']] + ' | '
+                if self._params['add_priority_to_case_header']:
+                    header_ending_list.append(self._priorities[case['priority_id']])
+                text = ''
+                if self._params['multi_param_name']:
+                    for item in case[self._params['multi_param_sys_name']]:
+                        text += self._params['multi_param_values'][item] + ', '
+                if self._params['add_multi_param_to_std_table']:
+                    table_row += text[:-2] + '  | '
+                if self._params['add_multi_param_to_case_header']:
+                    header_ending_list.append(text[:-2])
+                if self._params['add_case_id_to_std_table']:
+                    table_row += str(case['id']) + '  | '
+                if self._params['add_case_id_to_case_header']:
+                    header_ending_list.append('ID ' + str(case['id']))
+                table_row += case['title'] + ' |   |   |'
+                header_ending_list.reverse()
+                for item in header_ending_list:
+                    header_ending += item + '; '
+                header_ending = header_ending[:-2]
+                if header_ending:
+                    header_ending = ' \[' + header_ending + '\]'
+
+                self._std_table.append(table_row)
+                self._test_cases.append('##%s %s%s\n\n' % (title_level_up, case['title'], header_ending))
 
 # Test-case processing differs depending on the template id. All processors are in case_processing module.
                 case_template = '/'.join((str(self.project_path), self._template_folder, ''.join((str(case['template_id']), '.j2'))))
@@ -241,9 +334,9 @@ class Preprocessor(BasePreprocessor):
                 else:
                     try:
                         template = self._env.get_template(case_template)
-                        result = template.render(case=case, platform_name=self._platform_name).split('\r\n')
-                    except Exception:
-                        print(f"\n\nThere is problem with jinja template for test case template_id {case['template_id']} (case_id {case['id']}) in folder {self._template_folder}")
+                        result = template.render(case=case, params=self._params).split('\r\n')
+                    except Exception as exception:
+                        print(f"\nThere is problem with jinja template for test case template_id {case['template_id']} (case_id {case['id']}) in folder {self._template_folder}:\n{exception}")
                         if self._print_case_structure:
                             print('\nCase structure:')
                             pprint(case)
@@ -291,24 +384,21 @@ class Preprocessor(BasePreprocessor):
                     string_counter = 0
 
 
-    def _make_std_table_first_row(self, columns):
+    def _make_std_table_first_row(self, add_column_headers):
         first_row = '| '
         for i, header in enumerate(self._std_table_column_headers):
-            if i+1 in columns:
+            if add_column_headers[i]:
                 first_row += self._std_table_column_headers[i] + ' | '
         first_row = first_row.strip()
-
         return first_row
 
 
     def _std_table_aligning(self):
+        add_column_headers = [True, self._params['add_priority_to_std_table'], self._params['add_multi_param_to_std_table'], self._params['add_case_id_to_std_table'], True, True, True]
 
-        if self._add_case_id_to_case_name:
-            self._std_table_column_headers = self._make_std_table_first_row([1,2,3,4,5])
-            self._std_table = [self._std_table_column_headers] + self._std_table
-        else:
-            self._std_table_column_headers = self._make_std_table_first_row([1,3,4,5])
-            self._std_table = [self._std_table_column_headers] + self._std_table
+        self._std_table_column_headers = self._make_std_table_first_row(add_column_headers)
+
+        self._std_table = [self._std_table_column_headers] + self._std_table
 
         column_widths = [0 for i in range(self._std_table[0].count('|') - 1)]
         strings = []
@@ -348,10 +438,42 @@ class Preprocessor(BasePreprocessor):
     def _resolve_url(self):
         for i, string in enumerate(self._test_cases):
             if '![' in string:
-                if not self._platform_id == self.defaults['platform_id']:
-                    self._test_cases[i] = re.sub("(?<=[\]][\(])(\w*)", self._screenshots_url + "\g<1>" + '_' + self._platform_name + self._screenshots_ext, string)
+                if self._params['multi_param_select'][0] != '':
+                    self._test_cases[i] = re.sub("(?<=[\]][\(])(\w*)", self._screenshots_url + "\g<1>" + '_' + self._params['multi_param_select'][0] + self._screenshots_ext, string)
                 else:
                     self._test_cases[i] = re.sub("(?<=[\]][\(])(\w*)", self._screenshots_url + "\g<1>" + self._screenshots_ext, string)
+
+
+    def _get_multi_param_matches(self):
+        multi_param_matches = set()
+        self._params.update({'multi_param_values': {}})
+        values = []
+        testrail_params = self._client.send_get('get_case_fields')
+
+        for param in testrail_params:
+            if param['system_name'] == self._params['multi_param_sys_name']:
+                for item in param['configs']:
+                    if self._project_id in item['context']['project_ids']:
+                        values = item['options']['items'].split('\n')
+        for value in values:
+            if value.split(',')[1].strip().lower() in self._params['multi_param_select']:
+                multi_param_matches.add(int(value.split(',')[0].strip()))
+            self._params['multi_param_values'].update({int(value.split(',')[0].strip()): value.split(',')[1].strip()})
+
+        self._params.update({'multi_param_matches': multi_param_matches})
+
+
+    def _get_priority_matches(self):
+        priority_matches = set()
+        priorities = self._client.send_get('get_priorities')
+        self._priorities = {}
+
+        for priority in priorities:
+            if priority['name'].lower() in self._params['choose_priorities']:
+                priority_matches.add(priority['id'])
+            self._priorities.update({priority['id']: priority['name']})
+
+        self._params.update({'priority_matches': priority_matches})
 
 
     def apply(self):
@@ -360,6 +482,9 @@ class Preprocessor(BasePreprocessor):
         project_name = self._client.send_get('get_project/%s' % self._project_id)['name']
 
         self.logger.debug(f'Collect data from {self._testrail_url}, project {project_name}')
+
+        self._get_multi_param_matches()
+        self._get_priority_matches()
 
         project_suites = self._client.send_get('get_suites/%s' % self._project_id)